refactor(components.email-domain): move to webUniverseComponent
