+++ client/app/email-domain/email/account/change-password/email-domain-email-account-change-password.controller.js
@@ -7,14 +7,14 @@ angular.module('App').controller(
      * @param $stateParams
      * @param $translate
      * @param Alerter
-     * @param Emails
+     * @param WucEmails
      */
-    constructor($scope, $stateParams, $translate, Alerter, Emails) {
+    constructor($scope, $stateParams, $translate, Alerter, WucEmails) {
       this.$scope = $scope;
       this.$stateParams = $stateParams;
       this.$translate = $translate;
       this.Alerter = Alerter;
-      this.Emails = Emails;
+      this.WucEmails = WucEmails;
     }
 
     $onInit() {
@@ -56,12 +56,12 @@ angular.module('App').controller(
     changePasswordAccount() {
       let passwordPromise = null;
       if (_.get(this.$scope.currentActionData, 'delegate', false)) {
-        passwordPromise = this.Emails.changePasswordDelegatedAccount(
+        passwordPromise = this.WucEmails.changePasswordDelegatedAccount(
           this.currentAccount.email,
           this.model,
         );
       } else {
-        passwordPromise = this.Emails.changePasswordAccount(
+        passwordPromise = this.WucEmails.changePasswordAccount(
           this.$stateParams.productId,
           this.currentAccount.accountName,
           this.model,
