+++ client/app/email-domain/email/account/delete/email-domain-email-account-delete.js
@@ -7,14 +7,14 @@ angular.module('App').controller(
      * @param $stateParams
      * @param $translate
      * @param Alerter
-     * @param Emails
+     * @param WucEmails
      */
-    constructor($scope, $stateParams, $translate, Alerter, Emails) {
+    constructor($scope, $stateParams, $translate, Alerter, WucEmails) {
       this.$scope = $scope;
       this.$stateParams = $stateParams;
       this.$translate = $translate;
       this.Alerter = Alerter;
-      this.Emails = Emails;
+      this.WucEmails = WucEmails;
     }
 
     $onInit() {
@@ -23,7 +23,7 @@ angular.module('App').controller(
     }
 
     deleteAccount() {
-      this.Emails.deleteAccount(
+      this.WucEmails.deleteAccount(
         this.$stateParams.productId,
         this.currentAccount.accountName,
       )
