+++ client/app/email-domain/email/account/unlock/email-domain-email-account-unlock.controller.js
@@ -7,14 +7,14 @@ angular.module('App').controller(
      * @param $stateParams
      * @param $translate
      * @param Alerter
-     * @param Emails
+     * @param WucEmails
      */
-    constructor($scope, $stateParams, $translate, Alerter, Emails) {
+    constructor($scope, $stateParams, $translate, Alerter, WucEmails) {
       this.$scope = $scope;
       this.$stateParams = $stateParams;
       this.$translate = $translate;
       this.Alerter = Alerter;
-      this.Emails = Emails;
+      this.WucEmails = WucEmails;
     }
 
     $onInit() {
@@ -47,7 +47,7 @@ angular.module('App').controller(
     }
 
     changePasswordAccount() {
-      this.Emails.changePasswordAccount(
+      this.WucEmails.changePasswordAccount(
         this.$stateParams.productId,
         this.currentAccount.accountName,
         this.model,
