+++ client/app/email-domain/email/redirection/update/email-domain-email-redirection-update.controller.js
@@ -7,14 +7,14 @@ angular.module('App').controller(
      * @param $stateParams
      * @param $translate
      * @param Alerter
-     * @param Emails
+     * @param WucEmails
      */
-    constructor($scope, $stateParams, $translate, Alerter, Emails) {
+    constructor($scope, $stateParams, $translate, Alerter, WucEmails) {
       this.$scope = $scope;
       this.$stateParams = $stateParams;
       this.$translate = $translate;
       this.Alerter = Alerter;
-      this.Emails = Emails;
+      this.WucEmails = WucEmails;
     }
 
     $onInit() {
@@ -29,7 +29,7 @@ angular.module('App').controller(
     }
 
     updateRedirection() {
-      return this.Emails.updateRedirection(this.$stateParams.productId, {
+      return this.WucEmails.updateRedirection(this.$stateParams.productId, {
         id: this.model.id,
         data: {
           to: this.model.to,
